Create dashboard.html

+++ forms.py
@@ -1,5 +1,5 @@
 from flask_wtf import FlaskForm
-from wtforms import StringField, PasswordField, SubmitField, IntegerField
+from wtforms import StringField, PasswordField, SubmitField, IntegerField, DecimalField
 from wtforms.validators import DataRequired, Length, NumberRange
 
 class LoginForm(FlaskForm):
@@ -11,14 +11,14 @@ class AddCourseForm(FlaskForm):
     course_name = StringField('Name', validators=[DataRequired()])
     practice_of_standards_weight = IntegerField('Practice of Standards Weight', validators=[DataRequired(), NumberRange(min=0)])
     assessment_of_standards_weight = IntegerField('Assessment of Standards Weight', validators=[DataRequired(), NumberRange(min=0)])
-    practice_of_standards_grade = IntegerField('Practice of Standards Grade', validators=[DataRequired(), NumberRange(min=0)])
-    assessment_of_standards_grade = IntegerField('Assessment of Standards Grade', validators=[DataRequired(), NumberRange(min=0)])
+    practice_of_standards_grade = DecimalField('Practice of Standards Grade', validators=[DataRequired(), NumberRange(min=0)])
+    assessment_of_standards_grade = DecimalField('Assessment of Standards Grade', validators=[DataRequired(), NumberRange(min=0)])
     submit = SubmitField('Submit')
 
 class DeleteCourseForm(FlaskForm):
     submit = SubmitField('Delete')
 
 class UpdateCourseForm(FlaskForm):
-    new_grade = IntegerField('New Grade', validators=[DataRequired(), NumberRange(min=0)])
+    new_grade = DecimalField('New Grade', validators=[DataRequired(), NumberRange(min=0)])
     submit = SubmitField('Save')
